feat: api-handlers-decomposition-recovery-001 - Centralize shared HTTP helpers
